Check to see if snippet exists in YT object

+++ index.js
@@ -102,23 +102,24 @@ client.on('message', msg => {
 	  //console.log(msg.embeds[0])
 	  const ytslug = msg.content.match(/youtube\.com\/watch\?v=(.*)/i)[1]
 	  getYTtitle(ytslug).then(ytdata => {
-	  	console.log("YT title: ", ytdata.items[0].snippet.channelTitle)
-		const ytchannelname = ytdata.items[0].snippet.channelTitle
-		if(ytchannelname.includes("JobbytheHong") || ytchannelname.includes("JonTronShow")) {
-			var callout;
-			if(ytchannelname == "JobbytheHong") callout = "Jobby-posting"
-			if(ytchannelname == "JonTronShow") callout = "Jontron-posting"
-			const replaceEmbed = new MessageEmbed()
-			.setTitle(`${msg.author.username}, no ${callout}! Instead let\'s talk about ANYTHING else.`)
-			.setAuthor('Radbot')
-			.setDescription('')
-			.setThumbnail()
-			.setURL('https://tfwiki.net/wiki/Special:Random')
-			msg.channel.send(replaceEmbed).then(emsg => {
-				msg.delete()	
-			})
+		if(ytdata.items[0].hasOwnProperty('snippet')) {
+	  		console.log("YT title: ", ytdata.items[0].snippet.channelTitle)
+			const ytchannelname = ytdata.items[0].snippet.channelTitle
+			if(ytchannelname.includes("JobbytheHong") || ytchannelname.includes("JonTronShow")) {
+				var callout;
+				if(ytchannelname == "JobbytheHong") callout = "Jobby-posting"
+				if(ytchannelname == "JonTronShow") callout = "Jontron-posting"
+				const replaceEmbed = new MessageEmbed()
+				.setTitle(`${msg.author.username}, no ${callout}! Instead let\'s talk about ANYTHING else.`)
+				.setAuthor('Radbot')
+				.setDescription('')
+				.setThumbnail()
+				.setURL('https://tfwiki.net/wiki/Special:Random')
+				msg.channel.send(replaceEmbed).then(emsg => {
+					msg.delete()	
+				})
+			}
 		}
-		
 	  })
   }
 	
@@ -132,15 +133,15 @@ client.on('message', msg => {
       debug: false                 // is more verbose when set to true
     });
     
-    //if ( /flamewar/i.test(msg.content) ) {
-    //    msg.react("<:flamewar:691696266400235590>");
-    //}
-    //if ( /beast machines/i.test(msg.content) ) {
-    //    msg.react("<:yvonco:715978467728031814>");
-    //}
-    //if ( /^`*~*\_*\**bah\.?\,?\**\_*~*`*!*$/i.test(msg.content) ) {
-    //    msg.channel.send("Bah!");
-    //}
+    if ( /flamewar/i.test(msg.content) ) {
+        msg.react("<:flamewar:691696266400235590>");
+    }
+    if ( /beast machines/i.test(msg.content) ) {
+        msg.react("<:yvonco:715978467728031814>");
+    }
+    if ( /^`*~*\_*\**bah\.?\,?\**\_*~*`*!*$/i.test(msg.content) ) {
+        msg.channel.send("Bah!");
+    }
    
    if (/(thanks|thank you|big ups|poggers|coggers|based|thankee|themk|toggers),? \brad\b/i.test(msg.content)) {
        msg.react("<:aboutTheTransformers:656259059854344202>");
